* commit '4c94acb88ac2a1f26654a6ee18554e4895892c52':
  Remove LF in tar name
  Add log file to list
  Add test files
  Initial merge from master latest

+++ src/hdf/object/HObject.java
@@ -18,9 +18,8 @@ import java.io.Serializable;
 
 /**
  * The HObject class is the root class of all the HDF data objects. Every data
- * class has HObject as a superclass. All objects (Groups and Datasets)
- * implement the methods of this class. The following is the inherited structure
- * of HDF Objects.
+ * class has HObject as a superclass. All objects implement the methods of this
+ * class. The following is the inherited structure of HDF Objects.
  *
  * <pre>
  *                                 HObject
@@ -50,14 +49,14 @@ import java.io.Serializable;
  * should check the OID of the data object to avoid duplicate copies of the same
  * object.</b>
  * <p>
- * HDF4 objects are uniquely identified by the OID of the (tag_id, ref_id) pair.
+ * HDF4 objects are uniquely identified by the OID (tag_id, ref_id) pair.
  * The ref_id is the object reference count. The tag_id is a pre-defined number
  * to identify the type of object. For example, DFTAG_RI is for raster image,
  * DFTAG_SD is for scientific dataset, and DFTAG_VG is for Vgroup.
  * <p>
- * HDF5 objects are uniquely identified by the OID or object reference. The OID
- * is usually obtained by H5Rcreate(). The following example shows how to
- * retrieve an object ID from a file.
+ * HDF5 objects are uniquely identified by the OID containing just the object
+ * reference. The OID is usually obtained by H5Rcreate(). The following example
+ * shows how to retrieve an object ID from a file:
  *
  * <pre>
  * // retrieve the object ID
@@ -106,7 +105,7 @@ public abstract class HObject implements Serializable, DataFormat {
 
     /**
      * The full path of the data object. The full path always starts with the
-     * root, a slash. The path cannot be changed. Also, a path must ended with a
+     * root, a slash. The path cannot be changed. Also, a path must be ended with a
      * slash. For example, /arrays/ints/
      */
     private String             path;
@@ -370,7 +369,7 @@ public abstract class HObject implements Serializable, DataFormat {
      * Sets the path of the object.
      * <p>
      * setPath() is needed to change the path for an object when the name of a
-     * group conatining the object is changed by setName(). The path of the
+     * group containing the object is changed by setName(). The path of the
      * object in memory under this group should be updated to the new path to
      * the group. Unlike setName(), setPath() does not change anything in file.
      *
@@ -388,21 +387,21 @@ public abstract class HObject implements Serializable, DataFormat {
     }
 
     /**
-     * Opens an existing object such as dataset or group for access.
+     * Opens an existing object such as a dataset or group for access.
      *
      * The return value is an object identifier obtained by implementing classes
      * such as H5.H5Dopen(). This function is needed to allow other objects to
      * be able to access the object. For instance, H5File class uses the open()
-     * function to obtain object identifier for copyAttributes(int src_id, int
+     * function to obtain object identifier for copyAttributes(long src_id, long
      * dst_id) and other purposes. The open() function should be used in pair
-     * with close(int) function.
+     * with close(long) function.
      *
-     * @see hdf.object.HObject#close(int)
+     * @see hdf.object.HObject#close(long)
      *
      * @return the object identifier if successful; otherwise returns a negative
      *         value.
      */
-    public abstract int open();
+    public abstract long open();
 
     /**
      * Closes access to the object.
@@ -416,14 +415,14 @@ public abstract class HObject implements Serializable, DataFormat {
      * @param id
      *            The object identifier.
      */
-    public abstract void close(int id);
+    public abstract void close(long id);
 
     /**
      * Returns the file identifier of of the file containing the object.
      *
      * @return the file identifier of of the file containing the object.
      */
-    public final int getFID() {
+    public final long getFID() {
         if (fileFormat != null) {
             return fileFormat.getFID();
         }
@@ -484,7 +483,7 @@ public abstract class HObject implements Serializable, DataFormat {
     /**
      * Returns a cloned copy of the object identifier.
      * <p>
-     * The object OID cannot be modified once it is created. getIOD() clones the
+     * The object OID cannot be modified once it is created. getOID() clones the
      * object OID to ensure the object OID cannot be modified outside of this
      * class.
      *
