+++ src/hdf/object/fits/FitsGroup.java
@@ -113,14 +113,14 @@ public class FitsGroup extends Group
 
     // Implementing DataFormat
     @Override
-    public int open() {
+    public long open() {
         // not supported
         throw new UnsupportedOperationException("Unsupported operation for NetCDF.");
     }
 
     /** close group access */
     @Override
-    public void close(int gid) {
+    public void close(long gid) {
         // not supported
         throw new UnsupportedOperationException("Unsupported operation for NetCDF.");
     }
